server/requirements.txt
flask
flask-sqlalchemy
flask-session
flask-cors
flask-migrate
pymysql
python-dotenv
bcrypt
cryptography